third_party/WebKit/LayoutTests/inspector/console/console-command-copy-expected.txt

Tests that console's copy command is copying into front-end buffer.

InspectorFrontendHost.copyText: qwerty
InspectorFrontendHost.copyText: <p>
    Tests that console's copy command is copying into front-end buffer.
</p>
InspectorFrontendHost.copyText: {
  "foo": "bar"
}
InspectorFrontendHost.copyText: [object Object]
InspectorFrontendHost.copyText: {
  "0": {},
  "1": {},
  "2": {},
  "3": {},
  "4": {},
  "5": {},
  "6": {}
}
InspectorFrontendHost.copyText: NaN
InspectorFrontendHost.copyText: Infinity
InspectorFrontendHost.copyText: null
InspectorFrontendHost.copyText: undefined
InspectorFrontendHost.copyText: 1
InspectorFrontendHost.copyText: true
InspectorFrontendHost.copyText: false
InspectorFrontendHost.copyText: null
